+++ packages/shared/src/game-dsl/index.ts
@@ -1,0 +1,5 @@
+export type { GameDsl } from './GameDsl'
+export type { WorldDsl } from './GameDsl'
+export type { EntityDsl } from './GameDsl'
+export type { ComponentDsl } from './GameDsl'
+export { EMPTY_GAME_DSL } from './GameDsl'